modify: 디자인 변경으로 인한 회사정보입력 페이지 수정
