+++ hometerview/Sources/Presenter/UI/WriteCompanyInfo/EnterCompany/ViewModel/EnterCompanyViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  EnterCompanyViewModel.swift
+//  hometerview
+//
+//  Created by Ok Hyeon Kim on 2022/10/17.
+//
+
+import Foundation
+
+class EnterCompanyViewModel: BaseViewModel,
+                     ObservableObject {
+}
